backend/requirements.txt
absl-py
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
aiosmtplib
alembic
annotated-types
anyio
attrs
bcrypt
blinker
certifi
cffi
charset-normalizer
click
colorama
contourpy
cryptography
cycler
dlib
dnspython
ecdsa
email-validator
face-recognition
face_recognition_models
fastapi
fastapi-mail
flatbuffers
fonttools
frozenlist
greenlet
h11
httptools
idna
jax
jaxlib
Jinja2
kiwisolver
Mako
MarkupSafe
matplotlib
mediapipe
ml_dtypes
multidict
numpy
opencv-contrib-python
opencv-python
opencv-python-headless
opt_einsum
packaging
passlib
Pillow
propcache
protobuf
psycopg2-binary
pyasn1
pycparser
pydantic
pydantic-settings
pydantic_core
PyJWT
pyparsing
python-dateutil
python-dotenv
python-jose
python-magic
python-multipart
PyYAML
razorpay
requests
rsa
scipy
six
sniffio
sounddevice
SQLAlchemy
starlette
twilio
typing-inspection
typing_extensions
urllib3
uvicorn
watchfiles
websockets
yarl
# --- Added for Face Service ---
deepface